refactor autocomplete
